requirements.txt
azure-storage-blob==12.8.0
boto3==1.24.0
click==8.0.3
loguru==0.5.3
python-dotenv==0.19.2